chore(editor): tidy blocks exports + move the slash-menu UI into components
Follow-on cleanup to the extensions/ reorg:

- blocks.ts: collect the extension definitions under one `export { … }`
  barrel and hoist the module-level regex constants to the top of the file.
- Move the slash-command UI out of commandMenu into components.tsx, beside
  the other editor React components (CommandList, its types, filterCommands
  and the command set). commandMenu is now a JSX-free .ts module holding just
  the Tiptap Suggestion wiring, which imports the UI back.
- Drop dead code: an unused colorFor re-export (presence) and an unused
  opaqueId import + newClientId helper (hooks).
